Adapt for first use (#9)

+++ UnitTests/Integration/TestCaseTrait.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * Test Case for all of the integration tests.
- * php version 5.6
+ * php version 7.0
  *
  * @package WP_Syntex\Polylang_Phpunit\Integration
  */
@@ -15,12 +15,12 @@ use PLL_Admin_Model;
 use PLL_Install;
 use WP_Syntex\Polylang_Phpunit\TestCaseTrait as GlobalTestCaseTrait;
 use WP_UnitTest_Factory;
-use WP_UnitTestCase;
 
 /**
  * Test Case for all of the integration tests.
  */
 trait TestCaseTrait {
+
 	use GlobalTestCaseTrait;
 
 	/**
@@ -29,6 +29,14 @@ trait TestCaseTrait {
 	 * @var PLL_Admin_Model
 	 */
 	protected static $model;
+
+	/**
+	 * List of active plugins.
+	 * Array of paths to plugin files, relative to the plugins directory.
+	 *
+	 * @var array<non-falsy-string>
+	 */
+	protected $activePlugins = [];
 
 	/**
 	 * Initialization before all tests run.
@@ -62,11 +70,6 @@ trait TestCaseTrait {
 	public function set_up() {
 		parent::set_up();
 
-		/**
-		 * `$this->activePlugins` must be an array of paths to plugin files, relative to the plugins directory.
-		 *
-		 * @var array<string>
-		 */
 		if ( ! empty( $this->activePlugins ) ) {
 			add_filter( 'pre_option_active_plugins', [ $this, 'filterActivePlugins' ] );
 		}
@@ -150,18 +153,23 @@ trait TestCaseTrait {
 		}
 
 		// Delete the default categories first.
-		$terms = wp_get_object_terms( get_option( 'default_category' ), 'term_translations' );
+		$default_category = get_option( 'default_category' );
+		$default_category = is_numeric( $default_category ) ? (int) $default_category : 0;
 
-		if ( ! is_wp_error( $terms ) ) {
-			foreach ( $terms as $term ) {
-				wp_delete_term( $term->term_id, 'term_translations' ); // phpcs:ignore Squiz.NamingConventions.ValidVariableName.MemberNotCamelCaps
+		if ( $default_category ) {
+			$terms = wp_get_object_terms( $default_category, 'term_translations' );
+
+			if ( ! is_wp_error( $terms ) ) {
+				foreach ( $terms as $term ) {
+					wp_delete_term( $term->term_id, 'term_translations' ); // phpcs:ignore Squiz.NamingConventions.ValidVariableName.MemberNotCamelCaps
+				}
 			}
-		}
 
-		$terms = self::$model->term->get_translations( get_option( 'default_category' ) );
+			$terms = self::$model->term->get_translations( $default_category );
 
-		foreach ( $terms as $termId ) {
-			wp_delete_term( $termId, 'category' );
+			foreach ( $terms as $termId ) {
+				wp_delete_term( $termId, 'category' );
+			}
 		}
 
 		foreach ( $languages as $lang ) {
